feat(crap-core): #136 — move core orchestration + cli to crap-core (AnalysisOutput<P> shape moves)
S4 of the crap-core monorepo migration. Pure relocation + parameterization —
no new public APIs, no Reporter trait, no domain/ports changes.

## Moves

| From                                    | To                                  |
|-----------------------------------------|-------------------------------------|
| crates/crap4rs/src/core/mod.rs (1303 L) | crates/crap-core/src/core/mod.rs    |
| crates/crap4rs/src/cli/mod.rs           | crates/crap-core/src/cli/mod.rs     |
| crates/crap4rs/src/cli/delta_args.rs    | crates/crap-core/src/cli/delta_args.rs |
| crates/crap4rs/src/cli/view_args.rs     | crates/crap-core/src/cli/view_args.rs |

## Parameterization (ADR D9 mixed dispatch)

  pub fn analyze<P: ParseDiagnostic>(
      options: &AnalyzeOptions,
      complexity: &dyn ComplexityPort,
      coverage: &dyn CoveragePort<Diagnostic = P>,
  ) -> Result<AnalysisOutput<P>>

  pub fn run<P: ParseDiagnostic + std::fmt::Display>(
      cli: Cli,
      complexity: &dyn ComplexityPort,
      coverage: &dyn CoveragePort<Diagnostic = P>,
      tool_version: &str,
  ) -> ExitCode

  pub fn parse_args(tool_version: &str, long_version: &str) -> Cli

`AnalysisOutput<P>` now lives in crap-core/core (CAO B1, ADR D4 amended).
Generics on data containers; trait objects on port orchestration; tool
version threaded as &str so crap-core stays free of crap4rs-specific
env vars.

## crates/crap4rs/src/main.rs final shape (20 LOC)

  fn main() -> ExitCode {
      let cli = crap_core::cli::parse_args(env!("CARGO_PKG_VERSION"), env!("CRAP4RS_LONG_VERSION"));
      let src = cli.input.src.clone().unwrap_or_else(|| PathBuf::from("src"));
      let coverage = LcovParser::new(src.canonicalize().unwrap_or(src));
      let complexity = SynComplexityAdapter::new();
      crap_core::cli::run(cli, &complexity, &coverage, env!("CARGO_PKG_VERSION"))
  }

## Public API preservation

crap4rs/lib.rs becomes a comprehensive shim re-exporting v0.4.0 paths:
- crap4rs::core::analyze, AnalyzeOptions, AnalysisOutput (concrete to LcovParseDiagnostic)
- crap4rs::cli::{Cli, run, ...}
ADR D10 shim-module pattern. No version bump.

## Test relocation

18 analyze-using pipeline tests (real LCOV + syn fixtures) moved to
crates/crap4rs/tests/analyze_pipeline_tests.rs — they exercise full
orchestration through real adapters, which live in crap-core but are
wired by crap4rs. Language-agnostic tests (threshold resolver,
score_and_summarize, empty_passing_result) stay in crap-core/core.

## CI self-check (run-twice + LCOV path alignment)

.github/workflows/ci.yml self-check job:
- Adds "Verify LCOV paths workspace-relative" pre-step that greps SF
  records and fails on absolute paths (LCOV path alignment debug).
- Splits "Run crap4rs against itself" into two scoped runs:
    --src crates/crap-core/src
    --src crates/crap4rs/src
  Targeted per-crate boundary; `--src crates` would walk crap4ts/napi
  (post-S5), fixtures, proptest seeds.

## Wire-envelope canary

Content drifted (24413 → 10773 lines, expected: crap4rs/src lost cli).
Path-set diff verified empty (130 paths each). Shape preserved.

## Stats

Raw:        15 files changed, +3611 / -16883
Code-only:  excluding wire_envelope_snapshot drift, +1250 / -1473

## Acceptance gates verified

- cargo build --workspace clean
- cargo nextest run --workspace --all-targets: 906/906 passed
- cargo +1.88 check --workspace --all-targets clean
- cargo fmt --check + cargo clippy --workspace --all-targets -- -D warnings clean
- cargo deny check: advisories ok, bans ok, licenses ok, sources ok
- RUSTDOCFLAGS=-D warnings cargo doc --workspace --no-deps clean
- AST-purity: zero `^use syn|^use lcov` in domain/ + ports/
- Wire-envelope path-set diff: empty (130 paths)
- main.rs: 20 LOC
- analyze + cli::run signatures: <P: ParseDiagnostic>, &dyn ports
- --version: `crap4rs 0.4.0 (<sha> <date>)` (bin name preserved via runtime override)
- cargo mutants --check (in flight, 528+ ok, 0 problems at commit time)

## Follow-ups filed

- #150 — Make LcovParser path-strip aware after CLI/config merge
  (main.rs reads cli.input.src directly, bypassing config-file `src`)

Closes #136
