+++ .github/workflows/ci.yml
@@ -91,14 +91,27 @@ jobs:
         with:
           name: lcov
       - run: cargo build --release --package crap4rs
-      - name: Run crap4rs against itself
-        # Scoped to crates/crap4rs/src — crap-core is empty in S1, so a
-        # workspace-root --src would walk only the Rust adapter anyway.
-        # S2+ widens this to a per-crate run-twice once code lands in
-        # crap-core. Lifted `--exclude "cli/**"` (closes #109): cli/ now
-        # passes the gate after the prepare_pipeline / count_cognitive_expr
-        # split (#121, #122) and core::analyze decomposition (#57). Worst
-        # CRAP across the codebase is well under the default threshold of 25.
+      # S4 (#136) widens self-check from one-shot `--src crates/crap4rs/src`
+      # to per-crate run-twice. Both must pass. NOT `--src crates` —
+      # InputArgs.src is single-path; recursive walk would pull in
+      # crates/crap4ts/napi/ (post-S5), tests/fixtures/*.lcov, proptest
+      # seeds. Targeted `--src` per crate keeps the boundary strict.
+      - name: Verify LCOV paths workspace-relative
+        # If cargo llvm-cov emits absolute SF paths, LcovParser path
+        # normalization won't match WalkBuilder paths and self-CRAP
+        # silently produces zero matches.
+        run: |
+          set -euo pipefail
+          echo "First 5 SF records in lcov.info:"
+          head -200 lcov.info | grep '^SF:' | head -5
+          if head -200 lcov.info | grep -q '^SF:/'; then
+            echo "::error::lcov.info has absolute SF paths; self-CRAP path matching will fail"
+            echo "  hint: cargo llvm-cov needs --workspace-root . or --remap-path-prefix"
+            exit 1
+          fi
+      - name: Self-CRAP — crap-core
+        run: ./target/release/crap4rs --coverage lcov.info --src crates/crap-core/src --strict --color always
+      - name: Self-CRAP — crap4rs
         run: ./target/release/crap4rs --coverage lcov.info --src crates/crap4rs/src --strict --color always
 
   self-crap-view:
